requirements.txt
aioredis==1.3.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
awsebcli==3.20.2
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blessed==1.18.1
botocore==1.21.56
bs4==0.0.1
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
click==8.0.1
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
daphne==3.0.2
dateparser==1.0.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.2
django-cluster-redis==1.0.5
django-cors-headers==3.10.1
django-invitations==1.9.3
django-pwa==1.0.10
django-redis==5.0.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangochannelsrestframework==0.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
facebook-sdk==3.1.0
future==0.16.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
msgpack==1.0.2
msgpack-python==0.5.6
mypy-extensions==0.4.3
numpy==1.20.1
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
paramiko==2.7.2
parso==0.8.2
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.3.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pytesseract==0.3.8
PyJWT==2.0.1
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyrsistent==0.18.0
python-dateutil==2.8.1
python3-openid==3.2.0
python-twitter==3.5
pytz==2021.1
PyYAML==5.3.1
razorpay==1.2.0
redis==3.0.0
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
semantic-version==2.8.5
service-identity==21.1.0
setuptools-rust==0.12.1
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
supervisor==4.2.2
termcolor==1.1.0
texttable==1.6.4
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.1.9
websocket-client==0.59.0
zope.interface==5.4.0
